api/requirements.txt
alembic==1.15.2
aniso8601==10.0.1
blinker==1.9.0
click==8.1.8
colorama==0.4.6
Flask==3.1.1
flask-cors==6.0.0
Flask-Migrate==3.1.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
greenlet==3.2.2
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
pytz==2025.2
six==1.17.0
SQLAlchemy==2.0.41
typing_extensions==4.13.2
Werkzeug==3.1.3